requirements.txt
fastapi
uvicorn
python-nmap